// GestionStock/src/app/services/openapi/model/article-request.ts
/**
 * Gestion de stock REST API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ArticleRequest { 
    id?: number;
    code: string;
    designation: string;
    prixUnitaireHt: number;
    tauxTva: number;
    categorie: number;
}